feat: Move SecurityInfo to connection module
